move secured books  methods to BookRestController to make progam more neater
