+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-spring.datasource.url=jdbc:mariadb://192.168.0.117:3306/library
+spring.datasource.url=jdbc:mariadb://localhost:3306/library
 spring.datasource.username=root
 spring.datasource.password=root
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
